backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
sse-starlette==2.1.3
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langchain-core==0.3.28
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
